+++ filip/models/ngsi_v2/context.py
@@ -174,7 +174,7 @@ class ContextEntity(ContextEntityKeyValues):
     "ContextAttribute"-model. Obviously, id and type are
     not allowed to be used as attribute names.
 
-    Example:
+    Example::
 
         >>> data = {'id': 'MyId',
                     'type': 'MyType',