+++ filip/utils/validators.py
@@ -15,9 +15,11 @@ def validate_http_url(url: AnyHttpUrl) -> str:
     """
     Function checks whether the host has "http" added in case of http as
     protocol.
+
     Args:
         url (AnyHttpUrl): the url for the host / port
-    Returns:
+
+    #Returns:
         validated url
     """
     return url
@@ -27,8 +29,10 @@ def validate_http_url(url: AnyHttpUrl) -> str:
 def validate_mqtt_url(url: AnyMqttUrl) -> str:
     """
     Function that checks whether a url is valid mqtt endpoint
+
     Args:
         url: the url for the target endpoint
+
     Returns:
        validated url
     """