+++ tests/clients/test_ngsi_v2_iota.py
@@ -32,7 +32,7 @@ class TestAgent(unittest.TestCase):
             service_path=settings.FIWARE_SERVICEPATH)
         clear_all(fiware_header=self.fiware_header,
                   cb_url=settings.CB_URL,
-                  iota_url=settings.IOTA_URL)
+                  iota_url=settings.IOTA_JSON_URL)
         self.service_group1 = ServiceGroup(entity_type='Thing',
                                            resource='/iot/json',
                                            apikey=str(uuid4()))
@@ -53,12 +53,12 @@ class TestAgent(unittest.TestCase):
             "expressionLanguage": None
         }
         self.client = IoTAClient(
-            url=settings.IOTA_URL,
+            url=settings.IOTA_JSON_URL,
             fiware_header=self.fiware_header)
 
     def test_get_version(self):
         with IoTAClient(
-                url=settings.IOTA_URL,
+                url=settings.IOTA_JSON_URL,
                 fiware_header=self.fiware_header) as client:
             self.assertIsNotNone(client.get_version())
 
@@ -67,7 +67,7 @@ class TestAgent(unittest.TestCase):
 
     @clean_test(fiware_service=settings.FIWARE_SERVICE,
                 fiware_servicepath=settings.FIWARE_SERVICEPATH,
-                iota_url=settings.IOTA_URL)
+                iota_url=settings.IOTA_JSON_URL)
     def test_service_group_endpoints(self):
         self.client.post_groups(service_groups=[self.service_group1,
                                                 self.service_group2])
@@ -78,9 +78,6 @@ class TestAgent(unittest.TestCase):
         self.client.get_group(resource=self.service_group1.resource,
                               apikey=self.service_group1.apikey)
 
-        clear_all(fiware_header=self.fiware_header,
-                  iota_url=settings.IOTA_URL)
-
     def test_device_model(self):
         device = Device(**self.device)
         self.assertEqual(self.device,
@@ -89,19 +86,13 @@ class TestAgent(unittest.TestCase):
     @clean_test(fiware_service=settings.FIWARE_SERVICE,
                 fiware_servicepath=settings.FIWARE_SERVICEPATH,
                 cb_url=settings.CB_URL,
-                iota_url=settings.IOTA_URL)
+                iota_url=settings.IOTA_JSON_URL)
     def test_device_endpoints(self):
         """
         Test device creation
         """
-        # Clean up Fiware test state, this test can fail if the device was not
-        # correctly removed before
-        clear_all(fiware_header=self.fiware_header,
-                  cb_url=settings.CB_URL,
-                  iota_url=settings.IOTA_URL)
-
         with IoTAClient(
-                url=settings.IOTA_URL,
+                url=settings.IOTA_JSON_URL,
                 fiware_header=self.fiware_header) as client:
             client.get_device_list()
             device = Device(**self.device)
@@ -135,15 +126,11 @@ class TestAgent(unittest.TestCase):
             self.assertEqual(self.fiware_header.service_path,
                              device_res.service_path)
 
-            #cleanup
-            clear_all(fiware_header=self.fiware_header,
-                      cb_url=settings.CB_URL,
-                      iota_url=settings.IOTA_URL)
 
     @clean_test(fiware_service=settings.FIWARE_SERVICE,
                 fiware_servicepath=settings.FIWARE_SERVICEPATH,
                 cb_url=settings.CB_URL,
-                iota_url=settings.IOTA_URL)
+                iota_url=settings.IOTA_JSON_URL)
     def test_metadata(self):
         """
         Test for metadata works but the api of iot agent-json seems not
@@ -163,7 +150,7 @@ class TestAgent(unittest.TestCase):
         logger.info(device.json(indent=2))
 
         with IoTAClient(
-                url=settings.IOTA_URL,
+                url=settings.IOTA_JSON_URL,
                 fiware_header=self.fiware_header) as client:
             client.post_device(device=device)
             logger.info(client.get_device(device_id=device.device_id).json(
@@ -175,11 +162,6 @@ class TestAgent(unittest.TestCase):
             logger.info(client.get_entity(entity_id=device.entity_name).json(
                 indent=2))
 
-        #clean up
-        clear_all(fiware_header=self.fiware_header,
-                  cb_url=settings.CB_URL,
-                  iota_url=settings.IOTA_URL)
-
     def tearDown(self) -> None:
         """
         Cleanup test server
@@ -187,4 +169,4 @@ class TestAgent(unittest.TestCase):
         self.client.close()
         clear_all(fiware_header=self.fiware_header,
                   cb_url=settings.CB_URL,
-                  iota_url=settings.IOTA_URL)
+                  iota_url=settings.IOTA_JSON_URL)